workshops/microsoft-unstructured-bytewax/pipelines/ingestion-pipelines/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-storage-queue
azure-identity
bytewax==0.19
pandas==2.2.2
requests==2.31.0
python-dotenv
pyyaml
azure-ai-documentintelligence
azure-ai-formrecognizer
azure-search-documents
markdownify
haystack-ai
unstructured-fileconverter-haystack
unstructured-client
"pinecone-client[grpc]"
